Move previous and next buttons to separated components
