+++ packages/rpv-page-navigation/src/StoreProps.ts
@@ -1,0 +1,8 @@
+import { ViewerState } from '@phuocng/rpv';
+
+interface StoreProps {
+    jumpToPage?(pageIndex: number): void;
+    getViewerState?: () => ViewerState;
+}
+
+export default StoreProps;
