+++ packages/rpv/src/store/createStore.ts
@@ -3,14 +3,14 @@ type StoreState = Record<string, any>;
 type StoreKey<T extends StoreState> = string & keyof T;
 type StoreHandlerType<T> = (params: T) => void;
 
-interface Store<T extends StoreState> {
+interface StoreProps<T extends StoreState> {
     subscribe<K extends StoreKey<T>>(eventName: K, handler: StoreHandlerType<NonNullable<T[K]>>): void;
     unsubscribe<K extends StoreKey<T>>(eventName: K, handler: StoreHandlerType<NonNullable<T[K]>>): void;
     update<K extends StoreKey<T>>(eventName: K, params: T[K]): void;
     get<K extends StoreKey<T>>(eventName: K): T[K];
 }
 
-function createStore<T extends StoreState>(initialState?: T): Store<T> {
+function createStore<T extends StoreState>(initialState?: T): StoreProps<T> {
     let state: T = initialState || {} as T;
 
     const listeners: {
@@ -38,4 +38,5 @@ function createStore<T extends StoreState>(initialState?: T): Store<T> {
 };
 
 export type StoreHandler<T> = StoreHandlerType<T>;
+export type Store<T> = StoreProps<T>;
 export default createStore;
